+++ src/reducers/triggerReducer.js
@@ -47,14 +47,10 @@ const TRIGGER_SINGLE_SET = 'TRIGGER_SINGLE_SET';
 const TRIGGER_TOGGLE = 'TRIGGER_TOGGLE';
 const TRIGGER_VALUE_SET = 'TRIGGER_VALUE_SET';
 
-export const triggerLevelSetAction = level => {
-    if (level === null) return { type: TRIGGER_VALUE_SET, triggerLevel: null };
+export const triggerLevelSetAction = triggerLevel => {
     return {
         type: TRIGGER_VALUE_SET,
-        triggerLevel:
-            level > 1000
-                ? Math.round(level / 1000) * 1000
-                : Math.max(1, Math.round(level)),
+        triggerLevel,
     };
 };